Change being able to train other datasets
