+++ fanogan/train_encoder_izif.py
@@ -1,13 +1,7 @@
-import argparse
 import os
 import torch
 import torch.nn as nn
-from torch.utils.data import DataLoader
-import torchvision.transforms as transforms
 from torchvision.utils import save_image
-
-from model import Generator, Discriminator, Encoder
-from tools import SimpleDataset, load_mnist
 
 
 """
@@ -84,58 +78,3 @@ def train_encoder_izif(opt, generator, discriminator, encoder,
 
                 batches_done += opt.n_critic
     torch.save(encoder.state_dict(), "results/encoder")
-
-
-def main(opt):
-    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-    (x_train, y_train), _ = load_mnist("dataset",
-                                       training_label=opt.training_label,
-                                       split_rate=opt.split_rate)
-    train_mnist = SimpleDataset(x_train, y_train,
-                                transform=transforms.Compose(
-                                    [transforms.ToPILImage(),
-                                     transforms.ToTensor(),
-                                     transforms.Normalize([0.5], [0.5])])
-                                )
-    train_dataloader = DataLoader(train_mnist, batch_size=opt.batch_size,
-                                  shuffle=True)
-
-    img_shape = (opt.channels, opt.img_size, opt.img_size)
-    generator = Generator(img_shape, opt.latent_dim)
-    discriminator = Discriminator(img_shape)
-    encoder = Encoder(img_shape)
-
-    train_encoder_izif(opt, generator, discriminator, encoder,
-                       train_dataloader, device)
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--n_epochs", type=int, default=200,
-                        help="number of epochs of training")
-    parser.add_argument("--batch_size", type=int, default=64,
-                        help="size of the batches")
-    parser.add_argument("--lr", type=float, default=0.0002,
-                        help="adam: learning rate")
-    parser.add_argument("--b1", type=float, default=0.5,
-                        help="adam: decay of first order momentum of gradient")
-    parser.add_argument("--b2", type=float, default=0.999,
-                        help="adam: decay of first order momentum of gradient")
-    parser.add_argument("--latent_dim", type=int, default=100,
-                        help="dimensionality of the latent space")
-    parser.add_argument("--img_size", type=int, default=28,
-                        help="size of each image dimension")
-    parser.add_argument("--channels", type=int, default=1,
-                        help="number of image channels")
-    parser.add_argument("--n_critic", type=int, default=5,
-                        help="number of training steps for "
-                             "discriminator per iter")
-    parser.add_argument("--sample_interval", type=int, default=400,
-                        help="interval betwen image samples")
-    parser.add_argument("--training_label", type=int, default=0,
-                        help="label for normal images")
-    parser.add_argument("--split_rate", type=float, default=0.8,
-                        help="rate of split for normal training data")
-    opt = parser.parse_args()
-
-    main(opt)
